Add feature to fetch random image from reddit
The reddit provider searches the top items for the previous week. This
leads to the top image being fairly static. In order to get a better
selection of images, the provider now defaults to returning a random
image. To return the top image, the user can provide `--random=false`
at runtime.